requirements.txt
ffmpeg-python==0.2.0
future==0.18.3
pdf2image==1.16.3
Pillow==8.4.0
